+++ src/utils.ts
@@ -6,9 +6,16 @@ export function hasHighEnoughReScriptRelayVersion(version: string): boolean {
 }
 
 export function getPreferredFragmentPropName(onType: string): string {
-  return workspace
-    .getConfiguration("rescript-relay")
-    .get("preferShortNames") === true
-    ? onType.split(/(?=[A-Z])|_/g).pop() ?? onType
-    : onType;
+  const result =
+    workspace.getConfiguration("rescript-relay").get("preferShortNames") ===
+    true
+      ? onType.split(/(?=[A-Z])|_/g).pop() ?? onType
+      : onType;
+
+  // Handle common ReScript keywords
+  if (result.toLowerCase().endsWith("type")) {
+    return result.slice(0, result.length - 1);
+  }
+
+  return result;
 }